Add olm.substitutesFor annotation
When building the bundle, add the olm.substitutesFor annotation so that
rebuilds (via Freshmaker or otherwise) that have different build
versions appended to the end of the main version (e.g.
service-telemetry-operator-v1.5.15000000001-0.1234.p) can allow the
bundle to be upgraded without changing the skipRange, allowing for
rebuilds due to changes in the resulting base image, but no changes in
the actual operator itself can result in the latest bundle version being
deployed.

Related: RHOSOR-576