Narrow the scope of the smoke test
The intention of this change is to limit CI smoke testing to only our supported
components (ie. no ElasticSearch)

I didn't want to remove the code because we intend to support writing to an
external ES soon, and we might as well use all the same code for that validation.
When we get to that stage, we can just remove these conditionals.

In the meantime, this aligns our CI testing with the new default
observability_strategy so we're testing it the same way it would deploy.