Add bundle builds to stf-run-ci for index-based deployments
Initial working code to allow bundle builds to be created in support of
index-based (CatalogSource) deployments of STF using local builds.

Updates the create_builds.yml logic so that it allows deployments to
proceed when builds have already been created. If local builds are
enabled then if BuildConfigs have already been created, then the role
will lookup the latest Build object and set the internal image path so
that the deployment can continue. Primary function is for iterative
development.

Stubbed out functionality to start creating the index image is
available, but still needs to be developed fully.