Working development for index-based deployment
Completed the initial implementation of a local bundle build and
index-image created from bundle images. Generated via opm and loaded in
a CatalogSource that allows stf-run-ci to Subscribe to the
service-telemetry-operator package.

Created to allow testing of the properties.yaml now included in Service
Telemetry Operator to allow dependencies to be resolved without
pre-Subscribing to the Operators.

Closes: STF-1362