Updates stf-run-ci setup playbook to use cert-manager in stable-v1 channel
For newer OCP versions (OCP >= 4.12), cert-manager should be consumed
from the stable-v1 channel. This differs from previous versions of OCP,
in which we were consuming cert-manager from the tech-preview channel.

The way of installing both operators is different (the tech-preview
requires a specific namespace and operator group, whereas the stable-v1
version does not).

This change adds a conditional block checking the versions of OCP in which
we are running this playbook and executing the installation steps that
correspond which each version.