Fix infinite reconcile loop
Fix a potential infinite reconcile loop when deploying Grafana due to
updated hashing algorithm which changes the hash contents each time the
contents is hashed. Adds a fix which results in skipping the update of
the default-grafana-htpasswd Secret object to avoid the situation.

Also makes a small add to the when clause list for deploying
Elasticsearch so that only when it is enabled are the tasks included.

Closes: STF-1499
Signed-off-by: Leif Madsen <lmadsen@redhat.com>