Update grafana htpasswd management logic
Update the logic for grafana htpasswd management to no longer use sha256
hashes and instead use bcrypt. Update the logic so that we create the
initial contents of the secret before writing it to the first secret so
that we can avoid a second Secret k8s_info call. Create the secret if it
isn't already created. If it is created, then read the existing salt and
generate a hash with it. Update the secret and the annotation for
grafana deployment. If no changes are found, then the secret and the
annotation will not be updated, resulting in no grafana restart.

NOTE: this commit has debugging logic in it to show how it works.
Following commit removes it and won't be available in future squashed
history. See original pull-request.