Fix infinite reconcile loop (#445)
* Fix infinite reconcile loop

Fix a potential infinite reconcile loop when deploying Grafana due to
updated hashing algorithm which changes the hash contents each time the
contents is hashed. Adds a fix which results in skipping the update of
the default-grafana-htpasswd Secret object to avoid the situation.

Also makes a small add to the when clause list for deploying
Elasticsearch so that only when it is enabled are the tasks included.

Closes: STF-1499
Signed-off-by: Leif Madsen <lmadsen@redhat.com>

* Logic change so Secret updates on parameter change

Update the htpasswd secret when the ServiceTelemetry CRD is updated for
a basicAuth password change. Will work with existing deployments, adding
the annotation to an existing secret to get everything aligned.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>

* Syntax and lint fix

* Remove TODO for creating dashboards

Remove TODO for creating dashboards as management of dashboards is
outside the scope of STF.

* Update grafana htpasswd management logic

Update the logic for grafana htpasswd management to no longer use sha256
hashes and instead use bcrypt. Update the logic so that we create the
initial contents of the secret before writing it to the first secret so
that we can avoid a second Secret k8s_info call. Create the secret if it
isn't already created. If it is created, then read the existing salt and
generate a hash with it. Update the secret and the annotation for
grafana deployment. If no changes are found, then the secret and the
annotation will not be updated, resulting in no grafana restart.

NOTE: this commit has debugging logic in it to show how it works.
Following commit removes it and won't be available in future squashed
history. See original pull-request.

* Remove debug logic

* Adjust Ansible play names

---------

Signed-off-by: Leif Madsen <lmadsen@redhat.com>