Expose operator_sdk versions used in CI
Currently we use two different versions of the operator_sdk
in our CI. One is being used for building the operator bundles (v0.x)
and the other is being used for deploying with operator bundles (v1.x).

This change makes this fact more explicit across the code and also
enables the possibility of configuring which versions to use.