[stf-run-ci] Update pull_secret check for running bundles
When switching from tags to skipping using a bool, the check
``pull_secret is defined`` became invalid.
When a task is skipped, it still returns a value, which is captured with
`register: pull_secret`

The run bundle command was using `pull_secret is defined` to determine
whether to pass the pull-secret arg, and this check was defunct,
since pull_secret is always defined.

The solution for this is to re-set the value back to a 0-length
string, and change the conditional to check the length.

Additionally, the boolean values needed to be explicitly treated
as bools so that the registry setup tasks would be skipped