Remove logs API interface from STF (#483)
* Remove logs API interface from STF

The logs interface in STF has never been supported or productized. A
recent bug was found where Loki can be interfered with when deployed on
the same system as STF. Since this has never been documented or
productized, it is being removed, as OpenStack logging should be sent
off-cluster from rsyslog directly to Elasticsearch.

Closes STF-1504

* Remove logs interface from CRD

Missed removing the logs interface from the CRD and defaults.