Move the metrics_result retrieve
Move the metrics_result value retrieval to above an echo command so that
the result is of the curl command and not the echo command. Also change
the scripts to set +e so that the scripts don't exist right away since
there are checks at the end that set an appropriate exit code (and gives
more information since the script will complete vs exit immediately).

Reported by Chris in Slack