[stf-run-ci][create_catalog] Swap query for a command task (#470)
* [stf-run-ci][create_catalog] Swap query for a command task

Using query looks up the kubeconfig on localhost, rather than the host
that ansible is executing against. This behaviour is different from
either using the shell/command modules or using k8s modules.
For consistent behaviour, the queries are replaced with an alternative
way to get the same information that will have consistent behahaviour
whether executing against localhost or a remote host.