Add related images for Prometheus and Alertmanager
Add support for RELATED_IMAGES for other artifacts we need in a default
deployment when installing via disconnected.

These images are required to be specifically called out in disconnected
environments as the Prometheus and Alertmanager controllers in the
Cluster Observability Operator (COO) don't use the RELATED_IMAGE
environment variables specified in the downstream CSV for COO, so we
need to specify them ourselves and then deploy Prometheus and
Alertmanager using the spec.image reference in the Custom Resource.

Closes: JIRA#STF-1713