Use grafana_path_image by default
Remove the baseImage configuration for Grafana from the example CR so
that in the OLM UI (OCP console) the baseImage isn't a populated value
by default.

Use the value from grafana_image_path as the preferred default as it
will be populated from the RELATED_IMAGES_GRAFANA_IMAGE environment
variable, which then itself has a default of the base_image value in
defaults/main.yml. Only if the administrator sets a specific value for
baseImage should we deploy with that image.

We do this to allow for a better disconnected environment experience out
of the box.