Skip installing deps when installing from bundles
Dependencies for STF are being handled by OLM since
STF 1.5.3.

Related change https://github.com/infrawatch/service-telemetry-operator/pull/423

Except for Prometheus/COO and cert-manager (since
those are scoped cluster-wide and OLM cannot
automatically resolve conflicts)

Related change https://github.com/infrawatch/service-telemetry-operator/pull/472

So there is no need to pre subscribe to those when
deploying for index or bundles.

It is required, though, for local installs.

The only operator that was on this condition is
the AMQ Interconnect operator.
So this change updates the check to reflect this.