[zuul] Resolve install_yamls errors by skipping bootstrap tags
The ci-framework/playbooks/01-bootstrap.yml uses two tags bootstrap and
packages.
the bootstrap tag sets up repos and runs install_yamls for creating the
openstack operators. These are not needed for STF as we just want to use
the CRC setup

Also need to workaround the required install_yamls file not being
present (this will be addressed)