Create index image from pre-built bundles (#597)
* Create index image from pre-built bundles

Deploy from bundles can now be used along with
deploy from index enabled.

This way, an index image can be created with
specific pre built bundles and deploy STF with
those.

* Fix set_fact syntax for SGO/SGO bundle info dicts

* Add ImageStreams for STO and SGO

* Fix minor typo in STO/SGO bundle info dict

* Set correct STO and SGO bundles path

We need to use the internal registry address
to build the index image that will serve
as base for the catalog

* Drop build number from SGO and STO bundles

We only need the tag when creating the index image
Keeping this value will generate errors

* Get STO and SGO tags from the bundle path

* Fix typo on SGO and STO images path

* Update README with bundle + index instructions

Add more details on how to deploy from index with
pre built bundles

* Use stf_channel on the deploy_stf step

Define this value to stable-1.5 when deploying
from bundles and index, unstable (the default)
if bundles have been built locally.

Also, honor the sgo and sto bundles tags. Ideally,
we would update the sgo and sto image path to be
only the path (without the tag) and define the tag
separately. In this case, it is responsibility
from the user to make sure that the tag in the
image path and the tag provided in sto and sgo
image path match when deploying from externally
built bundles and index.

* Update missing reference to SGO bundles channel

* Update some logic paths to simplify the execution

* Honor the stf_channel variable

Nightly bundles use the "unstable" channel while
RH catalog bundles use "stable-1.5" channel

We should use nightly bundles "unstable" by default
and pass "stable-1.5" when deploying from RH catalog

* Get opm for getting bundles information

We need opm to get the bundle version when deploying
from index with pre-built bundles

Get opm following the same logic we use for operator-sdk

* Get the operator bundle version for STO and SGO

Inspect the provided bundles with opm render and
obtain the operator bundle version for each of them.

* Set a default version for opm

Let's stick to latest-4.14 for now, with the option
of updating to a different one if needed

* Fix typo when executing opm render

* Revert "Fix typo when executing opm render"

This reverts commit 1050227747d741b2d9f1c11885bd2eb91fb5ee73.

* Revert "Set a default version for opm"

This reverts commit d6a51df8b2b5883e19a29c00abd1b0d6286c53af.

* Revert "Get the operator bundle version for STO and SGO"

This reverts commit c1d9db36574f3ddc2fa46a5cd1a7eff379d9b24f.

* Revert "Get opm for getting bundles information"

This reverts commit 642df50f8171b5709638c407cbc552df49068f27.

* Get STO and SGO operator bundle version using oc

oc image info command provides operator bundle
version information. Use this to get the proper
version numbers when building the index image

* Get STO and SGO bundle versions from info

Parse the oc image info output to get
the STO and SGO bundle versions

* Add extra logic to avoid variables overwrite

sto_bundle_info and sgo_bundle_info is overriding
even though the step is skipped

Add an extra logic check to avoid variables
override