Get tag from the name instead of the metadata (#640)
* Get tag from the name instead of the metadata

Retrieve image tag from name instead of the
image metadata when deploying from index with pre-built
bundles.

* Add operator_bundle_tag to s*o_bundle_info

Keep track of the tag being used for a bundle
in the operator_bundle_tag var

This is required when deploying from index +
prebuilt bundles, since the tag of the bundle
and the version in the metadata might differ

* Set a default for STO/SGO path when local build

The STO/SGO image path needs a default when doing
a local build. When deploying from index + prebuilt
bundles we set this path when the bundles get
imported.

Add a conditional to avoid overwriting facts.