Add support for deploy from catalog
Add __deploy_from_catalog and __stf_catalog_index_image_path options
to stf-run-ci.

This new option allows to create a CatalogSource using a pre-built
index image and deploy STF from it.

This is particularly useful to test package metadata issues that would
be invisible when testing bundles but very apparent when installing
from the actual catalog.

Closes: OSPRH-19028